+++ lib/developer/widgets/my_game_card.dart
@@ -1,35 +1,80 @@
 import 'package:flutter/material.dart';
 import 'package:learnvironment/developer/edit_game.dart';
 import 'package:learnvironment/main_pages/widgets/tag.dart';
+import 'package:learnvironment/services/data_service.dart';
+import 'package:provider/provider.dart';
 
-class MyGameCard extends StatelessWidget {
+class MyGameCard extends StatefulWidget {
   final String imagePath;
   final String gameTitle;
   final List<String> tags;
   final String gameId;
+  final bool isPublic;
   final Future<void> Function(String gameId) loadGame;
 
-  const MyGameCard({
+  MyGameCard({
     super.key,
     required this.imagePath,
     required this.gameTitle,
     required this.tags,
     required this.gameId,
     required this.loadGame,
+    required this.isPublic,
   });
+
+  @override
+  State<MyGameCard> createState() => _MyGameCardState();
+}
+
+class _MyGameCardState extends State<MyGameCard> {
+  bool _isPublic = false;
+  bool _isLoading = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _isPublic = widget.isPublic;
+  }
+
+  Future<void> _togglePublicStatus() async {
+    DataService dataService = Provider.of<DataService>(context, listen: false);
+
+    setState(() {
+      _isLoading = true;
+      _isPublic = !_isPublic;
+    });
+
+    try {
+      await dataService.updateGamePublicStatus(gameId: widget.gameId, status: _isPublic);
+
+      print('Game status updated');
+    } catch (e) {
+      print('Error toggling public status: $e');
+      setState(() {
+        _isPublic = !_isPublic;
+      });
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(content: Text('Failed to update game status!')),
+        );
+      }
+    } finally {
+      setState(() {
+        _isLoading = false;
+      });
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.all(12.0),
       child: GestureDetector(
-        key: Key('gameCard_$gameId'),
-        onTap: () async => await loadGame(gameId),
+        key: Key('gameCard_${widget.gameId}'),
+        onTap: () async => await widget.loadGame(widget.gameId),
         child: Container(
           decoration: BoxDecoration(
-            color: Theme
-                .of(context)
-                .cardColor,
+            color: Theme.of(context).cardColor,
             borderRadius: BorderRadius.circular(10),
             boxShadow: [
               BoxShadow(
@@ -46,7 +91,7 @@ class MyGameCard extends StatelessWidget {
               ClipRRect(
                 borderRadius: BorderRadius.vertical(top: Radius.circular(10)),
                 child: Image.asset(
-                  imagePath,
+                  widget.imagePath,
                   width: double.infinity,
                   fit: BoxFit.cover,
                 ),
@@ -55,49 +100,53 @@ class MyGameCard extends StatelessWidget {
                 padding: const EdgeInsets.all(10),
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.center,
-                  // Center content
                   children: [
-                    Center( // Center the game title
-                      child: Text(
-                        gameTitle,
-                        style: const TextStyle(
-                          fontWeight: FontWeight.bold,
-                          fontSize: 16,
-                        ),
-                        overflow: TextOverflow.ellipsis,
+                    Text(
+                      widget.gameTitle,
+                      style: const TextStyle(
+                        fontWeight: FontWeight.bold,
+                        fontSize: 16,
                       ),
+                      overflow: TextOverflow.ellipsis,
                     ),
                     const SizedBox(height: 8),
                     LayoutBuilder(
                       builder: (context, constraints) {
                         List<Widget> tagWidgets = [];
-                        for (int i = 0; i < tags.length && i < 3; i++) {
-                          tagWidgets.add(TagWidget(tag: tags[i]));
+                        for (int i = 0; i < widget.tags.length && i < 3; i++) {
+                          tagWidgets.add(TagWidget(tag: widget.tags[i]));
                         }
-                        if (tags.length > 3) {
+                        if (widget.tags.length > 3) {
                           tagWidgets.add(
-                            TagWidget(tag: '+${tags.length - 3} more'),
+                            TagWidget(tag: '+${widget.tags.length - 3} more'),
                           );
                         }
-                        return Center( // Center the tags
-                          child: Wrap(
-                            spacing: 5,
-                            runSpacing: 5,
-                            children: tagWidgets,
+                        return Wrap(
+                          spacing: 5,
+                          runSpacing: 5,
+                          children: tagWidgets,
+                        );
+                      },
+                    ),
+                    const SizedBox(height: 10),
+                    IconButton(
+                      icon: Icon(Icons.edit),
+                      key: Key("edit"),
+                      tooltip: 'Edit',
+                      onPressed: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => EditGame(gameId: widget.gameId),
                           ),
                         );
                       },
                     ),
-                   const SizedBox(height: 10),
-                    Center(
-                      child: IconButton(
-                        icon: Icon(Icons.edit),
-                        key: Key("edit"),
-                        tooltip: 'Edit',
-                        onPressed: () {
-                          Navigator.push(context, MaterialPageRoute(builder: (context) => EditGame(gameId: gameId,)));
-                        },
-                      ),
+                    ElevatedButton(
+                      onPressed: _isLoading ? null : _togglePublicStatus,
+                      child: _isLoading
+                          ? CircularProgressIndicator()  // Loading state
+                          : Text(_isPublic ? 'Make Private' : 'Make Public'),
                     ),
                   ],
                 ),