+++ CHANGELOG.md
@@ -76,6 +76,8 @@
 - As a student I want the games to offer tips so that I can improve my learning [#5](https://github.com/LEIC-ES-2024-25/2LEIC13T3/issues/5)
 - As a user I want to be able to hear a sound when I select an option in a game so that I can easily know if I selected the correct answer [#62](https://github.com/LEIC-ES-2024-25/2LEIC13T3/issues/62)
 - As a game developer I want to be able to create a game so that others could play it. [#15](https://github.com/LEIC-ES-2024-25/2LEIC13T3/issues/15)
+- As a game developer I want to be able to dry run my game so that I can ensure it is bug-free. [#17](https://github.com/LEIC-ES-2024-25/2LEIC13T3/issues/17)
+- As a game developer I want to be able to add a description to my game so that users can understand what it is about before playing it. [#20](https://github.com/LEIC-ES-2024-25/2LEIC13T3/issues/20)
 
 ### Detailed description
 - Teachers can now remove students from their classes [#107](https://github.com/LEIC-ES-2024-25/2LEIC13T3/pull/107)
@@ -86,6 +88,7 @@
 - When signing up, the username cannot be the same as the one of an already exiting account [#114](https://github.com/LEIC-ES-2024-25/2LEIC13T3/pull/114)
 - As a teacher, after deleting a class or an assignment, the app goes back to the "Home Page" to ensure refresh [#115](https://github.com/LEIC-ES-2024-25/2LEIC13T3/pull/115)
 - Students cannot see their classes in teacher mode [#103](https://github.com/LEIC-ES-2024-25/2LEIC13T3/issues/103)
+- Developers can now dry run their games [#117](https://github.com/LEIC-ES-2024-25/2LEIC13T3/pull/117)
 
 ### Test User
 - email: up202307719@g.uporto.pt